% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAUtils.R
\name{DownloadFASTQ}
\alias{DownloadFASTQ}
\title{DownloadFASTQ
Downloads FASTQ using a list or a data.frame from RetrieveProjectMetadata}
\usage{
DownloadFASTQ(l, folder = "./")
}
\value{
None
}
\description{
DownloadFASTQ
Downloads FASTQ using a list or a data.frame from RetrieveProjectMetadata
}
\examples{
DownloadFASTQ(metadata)

}

