% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAUtils.R
\name{ParallelDownloadFASTQ}
\alias{ParallelDownloadFASTQ}
\title{ParallelDownloadFASTQ}
\usage{
ParallelDownloadFASTQ(df, folder = "./", n = 6)
}
\value{
None
}
\description{
Parallel Downloads FASTQ using a data.frame from RetrieveProjectMetadata
}
\examples{
ParallelDownloadFASTQ(metadata)

}

