% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAUtils.R
\name{RetrieveProjectMetadata}
\alias{RetrieveProjectMetadata}
\title{RetrieveProjectData}
\usage{
RetrieveProjectMetadata(PID)
}
\value{
data.frame of metadata and urls
}
\description{
This function retrieves all associated metadata and urls to an SRP id
}
\examples{
RetrieveProjectMetadata("SRPXXXXXX")

}

